+++ AutoVisualizer/Component/StableDiffusion/GHC_SD_GenerateSync.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Drawing;
 using System.IO;
+using AutoVisualizer.Utils;
 using Grasshopper.Kernel;
 using Rhino.Geometry;
 
@@ -12,6 +13,8 @@ namespace AutoVisualizer.Component.StableDiffusion
         private string prompt = "";
         private string negPrompt = "";
         private dynamic request_data;
+        private List<string> outputPath = new List<string>();
+        private List<Image> image = new List<Image>();
 
         /// <summary>
         /// Initializes a new instance of the MyComponent1 class.
@@ -40,6 +43,7 @@ namespace AutoVisualizer.Component.StableDiffusion
         protected override void RegisterOutputParams(GH_Component.GH_OutputParamManager pManager)
         {
             pManager.AddTextParameter("Path", "P", "Output path of tmp image file", GH_ParamAccess.list);
+            pManager.AddGenericParameter("Image", "I", "A bitmap object", GH_ParamAccess.list);
         }
 
         /// <summary>
@@ -114,28 +118,48 @@ namespace AutoVisualizer.Component.StableDiffusion
                     //var length = responseJSON.Length;
                     var images = responseJSON["images"];
 
-                    if (images.Count > 0)
+                    if (images.Count == 1)
                     {
                         base64ImageData = images[0].ToString();
-                    }
 
-                    //base64ImageData = (responseJSON["images"] != null && responseJSON["images"].Count > 0) ? responseJSON["images"][0] : null;
-
-
-                    if (!string.IsNullOrEmpty(base64ImageData))
-                    {
-                        byte[] imageBytes = Convert.FromBase64String(base64ImageData);
-                        using (MemoryStream stream = new MemoryStream(imageBytes))
+                        if (!string.IsNullOrEmpty(base64ImageData))
                         {
-                            Image image = Image.FromStream(stream);
-                            string outputPath = "C:\\tmp\\output.png";
-                            image.Save(outputPath);
-                            DA.SetData(0, outputPath);
-                            AddRuntimeMessage(GH_RuntimeMessageLevel.Remark, $"Image saved to {outputPath}");
+                            byte[] imageBytes = Convert.FromBase64String(base64ImageData);
+                            using (MemoryStream stream = new MemoryStream(imageBytes))
+                            {
+                                this.image.Add(Image.FromStream(stream));
+                                this.outputPath.Add("C:\\tmp\\output.png");
+                                this.image[0].Save(outputPath[0]);
 
-                            message = "Image generated!";
-                            UpdateMessage(message);
+                                DA.SetData(0, this.outputPath[0]);
+                                DA.SetData(1, this.image[0]);
+
+                                message = "Image generated!";
+                                UpdateMessage(message);
+                            }
                         }
+                    }
+                    else if (images.Count > 1)
+                    {
+                        for (int i = 0; i < images.Count; i++)
+                        {
+                            byte[] imageBytes = Convert.FromBase64String(images[i].ToString());
+                            using (MemoryStream stream = new MemoryStream(imageBytes))
+                            {
+                                this.image.Add(Image.FromStream(stream));
+                                this.outputPath.Add("C:\\tmp\\output" + i + ".png");
+                                this.image[i].Save(outputPath[i]);
+
+                                //DA.SetData(0, outputPath);
+                                //AddRuntimeMessage(GH_RuntimeMessageLevel.Remark, $"Image saved to {outputPath}");
+
+                                message = "Images generated!";
+                                UpdateMessage(message);
+                            }
+                        }
+
+                        DA.SetDataList(0, this.outputPath);
+                        DA.SetDataList(1, this.image);
                     }
                     else
                     {